Clean up sans tutorial changes
The default length for the individual spin startes in sesans mode
has changed, so there are now more flipper states to measure.
Additionally, it's useful to be able to log debug data, so the
tutorial now only compares against the INFO level of logging, so that
debug statements can be changed without failed unit tests.